+++ CHANGELOG.md
@@ -17,7 +17,7 @@ Note that this project **does not** adhere to [Semantic Versioning](https://semv
 
 ### Fixed
 
-- EndNote importer now respects the citation key preferences for unwanted characters.
+- EndNote and Refer importers now respect the citation key preferences for unwanted characters.
 
 ### Removed